+++ R/zzz.R
@@ -16,7 +16,7 @@
     packageStartupMessage( 
       paste0(" Package npsp: ", pkg.info["Title"], ",\n"),
       paste0(" version ", pkg.info["Version"], " (built on ", pkg.info["Date"], ").\n"),
-      " Copyright (C) R. Fernandez-Casal 2012-2017.\n",
+      " Copyright (C) R. Fernandez-Casal 2012-2018.\n",
       " Type `help(npsp)` for an overview of the package and\n",
       ' `demo(package = "npsp")` for the list of available demos.\n')
 }